// src/store-modules/library/library-store-module-action-type.ts
enum LIBRARY_STORE_MODULE_ACTION_TYPE {
    DOCUMENT_CREATE_MANY = 'documentCreateMany',
    DOCUMENT_CREATE_ONE = 'documentCreateOne',
    DOCUMENT_DELETE_ONE = 'documentDeleteOne',
    DOCUMENT_READ_ALL = 'documentReadAll',
    DOCUMENT_UPDATE_ONE = 'documentUpdateOne'
}

export default LIBRARY_STORE_MODULE_ACTION_TYPE;