// src/store-modules/library/library-store-module-mutation-type.ts
enum LIBRARY_STORE_MODULE_MUTATION_TYPE {
    DOCUMENT_SET_ALL = 'documentSetAll',
    FACET_CALCULATE_ALL = 'facetCalculateAll',
    FACET_VALUE_ADD_SELECTED = 'facetValueAddSelected',
    FACET_VALUE_REMOVE_SELECTED = 'facetValueRemoveSelected',
    NOOP = 'noop'
}

export default LIBRARY_STORE_MODULE_MUTATION_TYPE;