// src/store-modules/list/lists-store-module-action-type.ts
enum LIST_STORE_MODULE_ACTION_TYPE {
    LIST_CREATE_ONE = "listCreateOne",
    LIST_DELETE_ONE = "listDeleteOne",
    LIST_READ_ALL = "listReadAll",
    LIST_READ_ONE = "listReadOne",
    LIST_UPDATE_ONE = "listUpdateOne",
    LIST_UPDATE_MANY = "listUpdateMany",

    LIST_FOCUSED_ADD_DOCUMENTS = "listAddDocuments",
    LIST_FOCUSED_READ_ALL_DOCUMENTS = "listReadAllDocuments"
}

export default LIST_STORE_MODULE_ACTION_TYPE;